requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
asyncssh==2.13.0
attr==0.3.1
attrs==23.2.0
awscli==1.33.29
azure-core==1.30.1
azure-storage-blob==12.19.1
bertviz==1.4.0
bleach==5.0.1
bokeh==3.4.1
boto==2.49.0
boto3==1.34.69
botocore==1.34.147
boxing==0.1.4
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
contourpy==1.2.0
crashtest==0.4.1
cryptography==42.0.5
cycler==0.12.1
dask==2021.5.1
datashader==0.16.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.9
distributed==2021.5.1
distro==1.9.0
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
dnspython==2.6.1
docker==4.4.4
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.16
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
duckdb==1.1.3
dulwich==0.21.7
durations==0.3.3
elementpath==4.6.0
email_validator==2.1.2
enum34==1.1.10
esig==0.9.8.3
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
expiringdict==1.2.2
falcon==3.1.3
fastjsonschema==2.20.0
ffmpy==0.3.2
filelock==3.13.1
filetype==1.2.0
fire==0.7.0
fonttools==4.50.0
fsspec==2024.3.1
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.1.1
grpcio==1.62.1
h11==0.14.0
HeapDict==1.0.1
holoviews==1.18.3
htmlmin==0.1.12
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humansignal-drf-yasg==1.21.9
ijson==3.2.3
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
ipython==8.12.3
ipywidgets==8.1.2
isodate==0.6.1
jaraco.classes==3.4.0
jedi==0.19.1
jiter==0.7.0
jmespath==1.0.1
joblib==1.3.2
jsonpath-python==1.0.6
jsonschema==3.2.0
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
keyring==24.3.1
kiwisolver==1.4.5
kornia==0.6.10
lapx==0.5.10.post1
launchdarkly-server-sdk==7.5.0
lazy_loader==0.4
lightning-utilities==0.7.1
linkify-it-py==2.0.3
locket==1.0.0
lockfile==0.12.2
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
morph_kgc==2.8.0
mpmath==1.3.0
msgpack==1.0.4
multipledispatch==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
nltk==3.6.7
numpyencoder==0.3.0
openai==1.54.3
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
ordered-set==4.0.2
orjson==3.10.5
owlready2==0.46
packaging==23.2
pandas
pandocfilters==1.5.1
panel==1.4.2
param==2.1.0
parso==0.8.3
partd==1.3.0
pathtools==0.1.2
patsy==0.5.3
pexpect==4.9.0
Pillow==10.1.0
pipreqs==0.5.0
pkginfo==1.11.2
platformdirs==4.3.6
plotly==5.22.0
pluggy==1.3.0
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf
prts==1.0.0.3
psutil==5.9.8
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyct==0.5.0
pydantic
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.17.2
pynndescent==0.5.12
pyod==1.0.7
pyoxigraph==0.3.22
pyparsing==3.1.1
pyproject_hooks==1.2.0
pyRFC3339==1.1
pyrsistent==0.20.0                                                        
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.4
python-multipart==0.0.9
pytorch-lightning==1.9.3
pyts==0.12.0
pytz==2022.7.1
pyviz_comms==3.0.2
PyYAML==6.0.1
RapidFuzz==3.10.1
rdflib==7.0.0
redis==3.5.3
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.48
rotten-tomatoes-client==0.0.3
rq==1.10.1
rsa==4.7.2
ruamel.yaml==0.18.0
ruamel.yaml.clib==0.2.12
ruff==0.4.9
rules==2.2
s3transfer==0.10.1
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
semantic-version==2.10.0
semver==2.13.0
sentry-sdk==2.18.0
serpapi==0.1.5
setproctitle==1.3.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.13.5
stumpy==1.11.1
supervision==0.19.0
sympy==1.12
tblib==1.7.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.5.10
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torchmetrics==0.11.1
traitlets==5.14.2
transformers==4.41.2
trove-classifiers==2024.10.21.16
tsfresh==0.20.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
ua-parser==0.18.0
uc-micro-py==1.0.3
ujson==5.9.0
ultralytics==8.2.35
ultralytics-thop==2.0.0
ultralytics-yolov5==0.1.1
umap-learn==0.5.6
uritemplate==4.1.1
urllib3
user-agents==2.2.0
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.27.1
wandb==0.18.7
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.14.1
xarray==2024.5.0
xattr==1.1.0
xmljson==0.2.0
xyzservices==2024.4.0
yarg==0.1.9
zict==2.2.0
zipp==3.18.1
